Tools(Inline): useInline in dev
